feat: move dashboard to mine dashboard and add workspace navigation
- Changed sidebar navigation labels from "Dashboard" to "My Dashboard" and added "Lab Dashboard", "Workspace", and "Personal".
- Updated breadcrumb labels accordingly.
- Introduced a new "Lab Dashboard" section with a coming soon message.
- Created a new route for user-specific dashboard statistics under "/mine/dashboard".
- Refactored backend to support new dashboard statistics structure and updated API routes.
