+++ svc-gateway/internal/handler/stats.go
@@ -13,7 +13,7 @@ import (
 )
 
 type StatsService interface {
-	GetDashboardStats(ctx context.Context, userID uint) (*dto.DashboardStatsResponse, error)
+	GetMyDashboardStats(ctx context.Context, userID uint) (*dto.MyDashboardStatsResponse, error)
 }
 
 type StatsHandler struct {
@@ -24,18 +24,18 @@ func NewStatsHandler(statsService StatsService) *StatsHandler {
 	return &StatsHandler{statsService: statsService}
 }
 
-func (h *StatsHandler) GetDashboardStats(c *gin.Context) {
+func (h *StatsHandler) GetMyDashboardStats(c *gin.Context) {
 	userID := c.GetUint("user_id")
 	if userID == 0 {
-		slog.Warn("GetDashboardStats: missing user ID in context")
+		slog.Warn("GetMyDashboardStats: missing user ID in context")
 		c.JSON(http.StatusUnauthorized, utils.ErrorResponse(fmt.Errorf("common.unauthorized")))
 		return
 	}
 
-	resp, err := h.statsService.GetDashboardStats(c.Request.Context(), userID)
+	resp, err := h.statsService.GetMyDashboardStats(c.Request.Context(), userID)
 	if err != nil {
-		slog.Error("GetDashboardStats service error", "err", err)
-		c.JSON(http.StatusInternalServerError, utils.ErrorResponse(fmt.Errorf("service.get_dashboard_stats.failed")))
+		slog.Error("GetMyDashboardStats service error", "err", err)
+		c.JSON(http.StatusInternalServerError, utils.ErrorResponse(fmt.Errorf("service.get_my_dashboard_stats.failed")))
 		return
 	}
 	c.JSON(http.StatusOK, resp)
